+++ inc/Base/BaseController.php
@@ -65,7 +65,7 @@ class BaseController
 		$this->plugin_path = plugin_dir_path( $this->dirname_r( __FILE__, 2 ) );
 		$this->plugin_url = plugin_dir_url( $this->dirname_r( __FILE__, 2 ) );
 		$this->plugin = plugin_basename( $this->dirname_r( __FILE__, 3 ) ) . '/wp-last-modified-info.php';
-		$this->version = '1.7.0';
+		$this->version = '1.7.1';
 		$this->debug = false;
 		$this->name = 'WP Last Modified Info';
 	}